requirements.txt
# Core Django
Django==5.2.6
django-environ==0.11.2
python-dateutil==2.9.0

# Database & Cache
psycopg2-binary==2.9.10
redis==5.2.0

# Authentication & Security
django-allauth==65.3.0
django-csp==3.8
django-ratelimit==4.1.0

# Task Queue
celery[redis]==5.4.0
flower==2.0.1

# Storage
boto3==1.35.82
django-storages==1.14.4

# HTTP & APIs
requests==2.32.3
mercadopago==2.2.3
fiscalapi==4.0.270  # FiscalAPI SDK para CFDI

# Real-time
django-eventstream==5.3.2

# Cryptography
cryptography==43.0.3

# Monitoring
sentry-sdk[django]==2.19.2

# Development & Monitoring
django-debug-toolbar==4.4.6

# Office & Export
xlsxwriter==3.2.0